kits/hello_llm/actors/chatgpt_engine/requirements.txt
# Llm related
python-dotenv
langchain
langchain-community
langchain-openai
langchain-core

# Paranet
paranet_agent